+++ lsystems/lsystem.py
