+++ lsystems/newton.py
