+++ setup.py
@@ -1,10 +1,10 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
-  name = 'recart',
-  version = '0.1',
-  author = 'Maximilian Schlund',
-  license = 'MIT',
-  package_dir={'' : 'src'},
-  description='A small program to draw recursive art like the dragon curve as svg.'
-  )
+    name='lsystems',
+    version='0.2',
+    author='Maximilian Schlund',
+    license='MIT',
+    packages=find_packages(['lsystems']),
+    description='A small program to draw recursive art like the dragon curve as svg.'
+)
